requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langchain-classic>=0.1.0
langchain-core>=0.3.0
